// .devops/aws/src/imports/k8s.d.ts
export declare class KubeDeployment extends ApiObject {
  /**
   * Returns the apiVersion and kind for "io.k8s.api.apps.v1.Deployment"
   */
  static readonly GVK: GroupVersionKind;
  /**
   * Renders a Kubernetes manifest for "io.k8s.api.apps.v1.Deployment".
   *
   * This can be used to inline resource manifests inside other objects (e.g. as templates).
   *
   * @param props initialization props
   */
  static manifest(props?: KubeDeploymentProps): any;
  /**
   * Defines a "io.k8s.api.apps.v1.Deployment" API object
   * @param scope the scope in which to define this object
   * @param id a scope-local name for the object
   * @param props initialization props
   */
  constructor(scope: Construct, id: string, props?: KubeDeploymentProps);
  /**
   * Renders the object to Kubernetes JSON.
   */
  toJson(): any;
}
// ...
export interface KubeDeploymentProps {
  /**
   * Standard object's metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
   *
   * @schema io.k8s.api.apps.v1.Deployment#metadata
   */
  readonly metadata?: ObjectMeta;
  /**
   * Specification of the desired behavior of the Deployment.
   *
   * @schema io.k8s.api.apps.v1.Deployment#spec
   */
  readonly spec?: DeploymentSpec;
}
// ...
export interface DeploymentSpec {
  /**
   * Minimum number of seconds for which a newly created pod should be ready without any of its container crashing, for it to be considered available. Defaults to 0 (pod will be considered available as soon as it is ready)
   *
   * @default 0 (pod will be considered available as soon as it is ready)
   * @schema io.k8s.api.apps.v1.DeploymentSpec#minReadySeconds
   */
  readonly minReadySeconds?: number;
  /**
   * Indicates that the deployment is paused.
   *
   * @schema io.k8s.api.apps.v1.DeploymentSpec#paused
   */
  readonly paused?: boolean;
  /**
   * The maximum time in seconds for a deployment to make progress before it is considered to be failed. The deployment controller will continue to process failed deployments and a condition with a ProgressDeadlineExceeded reason will be surfaced in the deployment status. Note that progress will not be estimated during the time a deployment is paused. Defaults to 600s.
   *
   * @default 600s.
   * @schema io.k8s.api.apps.v1.DeploymentSpec#progressDeadlineSeconds
   */
  readonly progressDeadlineSeconds?: number;
  /**
   * Number of desired pods. This is a pointer to distinguish between explicit zero and not specified. Defaults to 1.
   *
   * @default 1.
   * @schema io.k8s.api.apps.v1.DeploymentSpec#replicas
   */
  readonly replicas?: number;
  /**
   * The number of old ReplicaSets to retain to allow rollback. This is a pointer to distinguish between explicit zero and not specified. Defaults to 10.
   *
   * @default 10.
   * @schema io.k8s.api.apps.v1.DeploymentSpec#revisionHistoryLimit
   */
  readonly revisionHistoryLimit?: number;
  /**
   * Label selector for pods. Existing ReplicaSets whose pods are selected by this will be the ones affected by this deployment. It must match the pod template's labels.
   *
   * @schema io.k8s.api.apps.v1.DeploymentSpec#selector
   */
  readonly selector: LabelSelector;
  /**
   * The deployment strategy to use to replace existing pods with new ones.
   *
   * @schema io.k8s.api.apps.v1.DeploymentSpec#strategy
   */
  readonly strategy?: DeploymentStrategy;
  /**
   * Template describes the pods that will be created.
   *
   * @schema io.k8s.api.apps.v1.DeploymentSpec#template
   */
  readonly template: PodTemplateSpec;
}
// ...
export declare class Quantity {
  readonly value: any;
  static fromString(value: string): Quantity;
  static fromNumber(value: number): Quantity;
  private constructor();
}
// ...
export interface ResourceRequirements {
  /**
   * Limits describes the maximum amount of compute resources allowed. More info: https://kubernetes.io/docs/concepts/configuration/manage-resources-containers/
   *
   * @schema io.k8s.api.core.v1.ResourceRequirements#limits
   */
  readonly limits?: {
    [key: string]: Quantity;
  };
  /**
   * Requests describes the minimum amount of compute resources required. If Requests is omitted for a container, it defaults to Limits if that is explicitly specified, otherwise to an implementation-defined value. More info: https://kubernetes.io/docs/concepts/configuration/manage-resources-containers/
   *
   * @schema io.k8s.api.core.v1.ResourceRequirements#requests
   */
  readonly requests?: {
    [key: string]: Quantity;
  };
}